Revert r248352:
>   Replace all known uses of ln in the build process with appropriate
>   install -l invocations via new INSTALL_LINK and INSTALL_SYMLINK
>   variables.

This breaks the assumption mergemaster and etcupdate that no new
features of tools will be required to run "make distribution" or it's
prerequisites.

Plotting a way forward will take some time and nothing in 9-STABLE
depends on INSTALL_LINK and INSTALL_SYMLINK yet so backing this out seem
like the right short-term approach.

PR:		misc/177055